--- a/hexl/include/hexl/util/check.hpp
+++ b/hexl/include/hexl/util/check.hpp
@@ -16,11 +16,11 @@
 /// error
 /// @param[in] cond A boolean indication the condition
 /// @param[in] expr The expression to be logged
-#define HEXL_CHECK(cond, expr)                                       \
-  if (!(cond)) {                                                     \
-    LOG(ERROR) << expr << " in fuction: " << __FUNCTION__            \
-               << " in file: " __FILE__ << " at line: " << __LINE__; \
-    throw std::runtime_error("Error. Check log output");             \
+#define HEXL_CHECK(cond, expr)                              \
+  if (!(cond)) {                                            \
+    LOG(ERROR) << expr << " in fuction: " << __FUNCTION__   \
+               << " in file: " __FILE__ << ":" << __LINE__; \
+    throw std::runtime_error("Error. Check log output");    \
   }
 
 /// @brief If input has an element >= bound, logs the expression and throws an