--- a/hexl/ntt/fwd-ntt-avx512.cpp
+++ b/hexl/ntt/fwd-ntt-avx512.cpp
@@ -75,6 +75,8 @@ void FwdButterfly(__m512i* X, __m512i* Y, __m512i W_op, __m512i W_precon,
     __m512i W_Y = _mm512_hexl_mullo_epi<BitShift>(W_op, *Y);
     T = _mm512_hexl_mullo_add_lo_epi<BitShift>(W_Y, Q, neg_modulus);
   } else if (BitShift == 64) {
+    // Perform approximate computation of Q, as described in page 7 of
+    // https://arxiv.org/pdf/2003.04510.pdf
     __m512i Q = _mm512_hexl_mulhi_approx_epi<BitShift>(W_precon, *Y);
     __m512i W_Y = _mm512_hexl_mullo_epi<BitShift>(W_op, *Y);
     // Compute T in range [0, 4q)