--- a/hexl/util/avx512-util.hpp
+++ b/hexl/util/avx512-util.hpp
@@ -78,23 +78,46 @@ inline __m512i _mm512_hexl_mulhi_epi<32>(__m512i x, __m512i y) {
 template <>
 inline __m512i _mm512_hexl_mulhi_epi<64>(__m512i x, __m512i y) {
   // https://stackoverflow.com/questions/28807341/simd-signed-with-unsigned-multiplication-for-64-bit-64-bit-to-128-bit
-  __m512i lomask = _mm512_set1_epi64(0x00000000ffffffff);
-  __m512i xh =
-      _mm512_shuffle_epi32(x, (_MM_PERM_ENUM)0xB1);  // x0l, x0h, x1l, x1h
-  __m512i yh =
-      _mm512_shuffle_epi32(y, (_MM_PERM_ENUM)0xB1);  // y0l, y0h, y1l, y1h
-  __m512i w0 = _mm512_mul_epu32(x, y);               // x0l*y0l, x1l*y1l
-  __m512i w1 = _mm512_mul_epu32(x, yh);              // x0l*y0h, x1l*y1h
-  __m512i w2 = _mm512_mul_epu32(xh, y);              // x0h*y0l, x1h*y0l
-  __m512i w3 = _mm512_mul_epu32(xh, yh);             // x0h*y0h, x1h*y1h
-  __m512i w0h = _mm512_srli_epi64(w0, 32);
-  __m512i s1 = _mm512_add_epi64(w1, w0h);
-  __m512i s1l = _mm512_and_si512(s1, lomask);
-  __m512i s1h = _mm512_srli_epi64(s1, 32);
-  __m512i s2 = _mm512_add_epi64(w2, s1l);
-  __m512i s2h = _mm512_srli_epi64(s2, 32);
-  __m512i hi1 = _mm512_add_epi64(w3, s1h);
-  return _mm512_add_epi64(hi1, s2h);
+  __m512i lo_mask = _mm512_set1_epi64(0x00000000ffffffff);
+  // Shuffle high bits with low bits in each 64-bit integer =>
+  // x0_lo, x0_hi, x1_lo, x1_hi, x2_lo, x2_hi, ...
+  __m512i x_hi = _mm512_shuffle_epi32(x, (_MM_PERM_ENUM)0xB1);
+  // y0_lo, y0_hi, y1_lo, y1_hi, y2_lo, y2_hi, ...
+  __m512i y_hi = _mm512_shuffle_epi32(y, (_MM_PERM_ENUM)0xB1);
+  __m512i z_lo_lo = _mm512_mul_epu32(x, y);        // x_lo * y_lo
+  __m512i z_lo_hi = _mm512_mul_epu32(x, y_hi);     // x_lo * y_hi
+  __m512i z_hi_lo = _mm512_mul_epu32(x_hi, y);     // x_hi * y_lo
+  __m512i z_hi_hi = _mm512_mul_epu32(x_hi, y_hi);  // x_hi * y_hi
+
+  //                   x_hi | x_lo
+  // x                 y_hi | y_lo
+  // ------------------------------
+  //                  [x_lo * y_lo]    // z_lo_lo
+  // +           [z_lo * y_hi]         // z_lo_hi
+  // +           [x_hi * y_lo]         // z_hi_lo
+  // +    [x_hi * y_hi]                // z_hi_hi
+  //     ^-----------^ <-- only bits needed
+  //  sum_|  hi | mid | lo  |
+
+  // Low bits of z_lo_lo are not needed
+  __m512i z_lo_lo_shift = _mm512_srli_epi64(z_lo_lo, 32);
+
+  //                   [x_lo  *  y_lo] // z_lo_lo
+  //          + [z_lo  *  y_hi]        // z_lo_hi
+  //          ------------------------
+  //            |    sum_tmp   |
+  //            |sum_mid|sum_lo|
+  __m512i sum_tmp = _mm512_add_epi64(z_lo_hi, z_lo_lo_shift);
+  __m512i sum_lo = _mm512_and_si512(sum_tmp, lo_mask);
+  __m512i sum_mid = _mm512_srli_epi64(sum_tmp, 32);
+  //            |       |sum_lo|
+  //          + [x_hi   *  y_lo]       // z_hi_lo
+  //          ------------------
+  //            [   sum_mid2   ]
+  __m512i sum_mid2 = _mm512_add_epi64(z_hi_lo, sum_lo);
+  __m512i sum_mid2_hi = _mm512_srli_epi64(sum_mid2, 32);
+  __m512i sum_hi = _mm512_add_epi64(z_hi_hi, sum_mid);
+  return _mm512_add_epi64(sum_hi, sum_mid2_hi);
 }
 
 #ifdef HEXL_HAS_AVX512IFMA
@@ -125,22 +148,35 @@ template <>
 inline __m512i _mm512_hexl_mulhi_approx_epi<64>(__m512i x, __m512i y) {
   // https://stackoverflow.com/questions/28807341/simd-signed-with-unsigned-multiplication-for-64-bit-64-bit-to-128-bit
   __m512i lo_mask = _mm512_set1_epi64(0x00000000ffffffff);
-  __m512i xh =
-      _mm512_shuffle_epi32(x, (_MM_PERM_ENUM)0xB1);  // x0l, x0h, x1l, x1h
-  __m512i yh =
-      _mm512_shuffle_epi32(y, (_MM_PERM_ENUM)0xB1);  // y0l, y0h, y1l, y1h
-  // __m512i w0 = _mm512_mul_epu32(x, y);               // x0l*y0l, x1l*y1l
-  __m512i w1 = _mm512_mul_epu32(x, yh);   // x0l*y0h, x1l*y1h
-  __m512i w2 = _mm512_mul_epu32(xh, y);   // x0h*y0l, x1h*y0l
-  __m512i w3 = _mm512_mul_epu32(xh, yh);  // x0h*y0h, x1h*y1h
-  // __m512i w0h = _mm512_srli_epi64(w0, 32);
-  __m512i s1 = w1;  // _mm512_add_epi64(w1, w0h);
-  __m512i s1l = _mm512_and_si512(s1, lo_mask);
-  __m512i s1h = _mm512_srli_epi64(s1, 32);
-  __m512i s2 = _mm512_add_epi64(w2, s1l);
-  __m512i s2h = _mm512_srli_epi64(s2, 32);
-  __m512i hi1 = _mm512_add_epi64(w3, s1h);
-  return _mm512_add_epi64(hi1, s2h);
+  // Shuffle high bits with low bits in each 64-bit integer =>
+  // x0_lo, x0_hi, x1_lo, x1_hi, x2_lo, x2_hi, ...
+  __m512i x_hi = _mm512_shuffle_epi32(x, (_MM_PERM_ENUM)0xB1);
+  // y0_lo, y0_hi, y1_lo, y1_hi, y2_lo, y2_hi, ...
+  __m512i y_hi = _mm512_shuffle_epi32(y, (_MM_PERM_ENUM)0xB1);
+  __m512i z_lo_hi = _mm512_mul_epu32(x, y_hi);     // x_lo * y_hi
+  __m512i z_hi_lo = _mm512_mul_epu32(x_hi, y);     // x_hi * y_lo
+  __m512i z_hi_hi = _mm512_mul_epu32(x_hi, y_hi);  // x_hi * y_hi
+
+  //                   x_hi | x_lo
+  // x                 y_hi | y_lo
+  // ------------------------------
+  //                  [x_lo * y_lo]    // unused, resulting in approximation
+  // +           [z_lo * y_hi]         // z_lo_hi
+  // +           [x_hi * y_lo]         // z_hi_lo
+  // +    [x_hi * y_hi]                // z_hi_hi
+  //     ^-----------^ <-- only bits needed
+  //  sum_|  hi | mid | lo  |
+
+  __m512i sum_lo = _mm512_and_si512(z_lo_hi, lo_mask);
+  __m512i sum_mid = _mm512_srli_epi64(z_lo_hi, 32);
+  //            |       |sum_lo|
+  //          + [x_hi   *  y_lo]       // z_hi_lo
+  //          ------------------
+  //            [   sum_mid2   ]
+  __m512i sum_mid2 = _mm512_add_epi64(z_hi_lo, sum_lo);
+  __m512i sum_mid2_hi = _mm512_srli_epi64(sum_mid2, 32);
+  __m512i sum_hi = _mm512_add_epi64(z_hi_hi, sum_mid);
+  return _mm512_add_epi64(sum_hi, sum_mid2_hi);
 }
 
 #ifdef HEXL_HAS_AVX512IFMA