… test/test-ntt.cpp
… test/test-util-avx512.hpp
… test/test-util.hpp