Optimizes 64-bit NTT by 1-2% using approximate AVX512 mulhi computation, as described on page 7 of https://arxiv.org/pdf/2003.04510.pdf. No speedup in EltwiseMultMod observed. EltwiseReduceMod optimization to be done in separate PR.

On ICX with clang-12:
Before
```
BM_FwdNTT_AVX512DQ_64/1024/1        2.67 us         2.65 us      1583566
BM_FwdNTT_AVX512DQ_64/1024/4        2.59 us         2.58 us      1630653
BM_FwdNTT_AVX512DQ_64/4096/1        12.6 us         12.6 us       333106
BM_FwdNTT_AVX512DQ_64/4096/4        12.3 us         12.3 us       341457
BM_FwdNTT_AVX512DQ_64/16384/1       58.4 us         58.3 us        72051
BM_FwdNTT_AVX512DQ_64/16384/4       57.2 us         57.1 us        73604
```

After
```
BM_FwdNTT_AVX512DQ_64/1024/1        2.60 us         2.60 us      1618281
BM_FwdNTT_AVX512DQ_64/1024/4        2.58 us         2.58 us      1666235
BM_FwdNTT_AVX512DQ_64/4096/1        12.4 us         12.3 us       340243
BM_FwdNTT_AVX512DQ_64/4096/4        12.2 us         12.2 us       348965
BM_FwdNTT_AVX512DQ_64/16384/1       57.4 us         57.4 us        73218
BM_FwdNTT_AVX512DQ_64/16384/4       56.2 us         56.2 us        74844
```

Before
```
BM_InvNTT_AVX512DQ_64/1024/1        3.11 us         3.11 us      1350732
BM_InvNTT_AVX512DQ_64/1024/2        3.08 us         3.07 us      1366420
BM_InvNTT_AVX512DQ_64/4096/1        14.6 us         14.6 us       288524
BM_InvNTT_AVX512DQ_64/4096/2        14.4 us         14.4 us       292415
BM_InvNTT_AVX512DQ_64/16384/1       67.2 us         67.1 us        62648
BM_InvNTT_AVX512DQ_64/16384/2       66.4 us         66.4 us        63351
```

After
```
BM_InvNTT_AVX512DQ_64/1024/1        3.04 us         3.03 us      1384194
BM_InvNTT_AVX512DQ_64/1024/2        3.00 us         2.99 us      1406817
BM_InvNTT_AVX512DQ_64/4096/1        14.1 us         14.0 us       298983
BM_InvNTT_AVX512DQ_64/4096/2        13.9 us         13.9 us       302398
BM_InvNTT_AVX512DQ_64/16384/1       65.0 us         64.8 us        64723
BM_InvNTT_AVX512DQ_64/16384/2       64.4 us         64.3 us        65369
```